% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haddockWts.R
\docType{data}
\name{haddockWts}
\alias{haddockWts}
\title{Short-term forecast individual weights for had.27.46a20}
\format{
list
}
\source{
WGNSSK 2024
}
\usage{
data(haddockWts)
}
\description{
The \code{haddockWts} data contains weights at age for catch,
  landings, discards and stock as modeled by a cohort-based linear model
  approach (Jarowski, 2011). Values are provided by the stock assessor in
  text file to WGMIXFISH. Example data is from the assessment conducted in
  2024.

\itemize{
  \item catch - individual catch weight
  \item landings - individual landings weight
  \item discards - individual discards weight
  \item stock - individual stock weight
}
}
\examples{

data(haddockWts)
haddockWts


}
\keyword{datasets}
