% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSeffort.R
\docType{data}
\name{NSeffort}
\alias{NSeffort}
\title{Data.frame Contains information on the total effort by
  year and gear categories}
\format{
data.frame
}
\source{
Intercatch
}
\usage{
data(NSeffort)
}
\description{
The \code{NSeffort} data.frame provides example data for
  use in `plot_FOgearType`.


\itemize{
  \item ID
  \item Country
  \item Year
  \item Quarter
  \item Metier
  \item gear
  \item target
  \item mesh
  \item gearCat
  \item metCat
  \item Vessel_length
  \item FDF
  \item Area
  \item kw_days
  \item Days_at_sea
  \item No_vessels
}
}
\examples{

data(NSlandings)
head(NSlandings)


}
\keyword{datasets}
