% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSlandings.R
\docType{data}
\name{NSlandings}
\alias{NSlandings}
\title{Data.frame Contains information on the total landings by
  year and gear categories}
\format{
data.frame
}
\source{
Intercatch
}
\usage{
data(NSlandings)
}
\description{
The \code{NSlandings} data.frame provides example data for
  use in `plot_FOgearType` and `plot_catchCompSpp`.

\itemize{
  \item ID
  \item Country
  \item Year
  \item Quarter
  \item Metier
  \item gear
  \item target
  \item mesh
  \item gearCat
  \item metCat
  \item Vessel_length
  \item FDF
  \item Area
  \item Species
  \item MIXFISHspp
  \item Stock
  \item Landings
  \item Value
}
}
\examples{

data(NSlandings)
head(NSlandings)


}
\keyword{datasets}
