% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{plot_alluvial}
\alias{plot_alluvial}
\title{Alluvial plot}
\usage{
plot_alluvial(
  data,
  refTable,
  group_vars = c("fleet", "metier", "stock"),
  fill_var = c("stock"),
  group_labs = NULL,
  text_repel = FALSE,
  text_size = 2,
  mult_x = c(0.1, 0.1),
  nudge_x = 1/3,
  stratum_width = 1/3,
  stratum_col = "white",
  xlab = NULL,
  ylab = "Catch [t]",
  fillLegendTitle = "Stock",
  addLegend = TRUE,
  plotTitle = NULL
)
}
\arguments{
\item{data}{data.frame Contains information on for making an alluvial-type
plot (e.g. catch or landings). The data.frame should contain several
variables for grouping (e.g. `fleet`, `metier`, `stock`) as well as a
variable to plot values (`value`).}

\item{refTable}{data.frame Contains stock look-up information for consistent
plotting of colors. The data.frame should contain a variable that matches
the `fill_var` definition (Default: `fill_var = "stock"`), and also found
in the `data` object. Contains a variable `col` that defines the color
used to fill bars in plot. `order` defines the order of levels used in
plot.}

\item{group_vars}{vector. Contains variable names (order from left to right)
that define the alluvial axes
(Default: `group_vars = c("fleet", "metier", "stock")`)}

\item{fill_var}{character. Data variable name used for coloring alluvial
(Default: `fill_var = "stock"`).}

\item{group_labs}{vector Contains the x-axis labels used for the group_vars.
(Default: group_labs = group_vars).}

\item{text_repel}{logical Should stratum labels be repelled to prevent
overlapping (using `ggrepel::geom_text_repel`)
(Default: `text_repel = FALSE`)}

\item{text_size}{numeric Value for text label size
(Default: `text_size = 2`)}

\item{mult_x}{vector. Two values defining the multiplier to expand the
x-axis (Default: `mult_x = c(0.1, 0.1)`). Passes values to
`expansion(mult = mult_x)`.}

\item{nudge_x}{numeric. Value to nudge direction of stratum labels when
`text_repel = TRUE`. (Default: `nudge_x = 1/3`)}

\item{stratum_width}{numeric. Width of the stratum bar
(Default: `stratum_width = 1/3`)}

\item{stratum_col}{color definition. Color used to fll the non-stock strata
(Default: `stratum_col = "white"`)}

\item{xlab}{character X-axis label (Default: `xlab = NULL` removes label)}

\item{ylab}{character Y-axis label (Default: `ylab = Catch [t]`)}

\item{fillLegendTitle}{character Fill legend title
(Default: `fillLegendTitle = "Stock"`)}

\item{addLegend}{logical Should legend of stock fill colors be added}

\item{plotTitle}{character Plot title
(Default: `plotTitle = NULL` removes title)}
}
\value{
plot output of class ggplot
}
\description{
Alluvial plot
}
\examples{

data("stfMtStkSum")
data("refTable")

## create catch data by fleet, metier, and stock
dataYr <- 2020
# add country for filter (if desired)
stfMtStkSum$country <- unlist(lapply(strsplit(stfMtStkSum$fleet, "_"),
  function(x){x[1]}))
# filter data
df <- subset(stfMtStkSum, year == 2020 & scenario == "min" &
  country == "BE")[,c("fleet", "metier", "stock", "landings")]
df$stock <- refTable$stock[match(df$stock, refTable$stock_short)]
names(df)[4] <- "value"

plot_alluvial(data = df, refTable = refTable,
  group_vars = c("fleet", "metier", "stock"),
  fill_var = c("stock"))

# doesn't make sense, but redefine group_var order
plot_alluvial(data = df, refTable = refTable,
  group_vars = c("fleet", "stock", "metier"),
  fill_var = c("stock"))

# a 2-level version
plot_alluvial(data = df, refTable = refTable,
  group_vars = c("fleet", "stock"),
  fill_var = c("stock"))

# repel labels
plot_alluvial(data = df, refTable = refTable,
  group_vars = c("fleet", "metier", "stock"),
  text_repel = TRUE, text_size = 2)

# repel labels and suppress legend
plot_alluvial(data = df, refTable = refTable,
  text_repel = TRUE, text_size = 2, addLegend = FALSE)

}
