% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_change.R
\name{plot_catch_change}
\alias{plot_catch_change}
\title{Plot fleet landings taken up relative to recent landings / quota}
\usage{
plot_catch_change(
  data = NULL,
  basis = "recent_catch",
  dataYrs = NULL,
  advYr = NULL,
  sc = "min",
  fleets_excl = NULL,
  refTable = NULL,
  xlab = "Stock",
  ylab = "catch change (tonnes)",
  fillLegendTitle = "Stock",
  colLegendTitle = "Limiting stock"
)
}
\arguments{
\item{data}{data.frame Contains information on catch by fleet and stock}

\item{basis}{is a character vector with the basis on which to compare the
scenario landings, either 'recent_catch' or 'Quota'. When 'recent_catch' is
used, the average landings from the defined years (argument `dataYrs`) is used as the
reference instead of the advice year quota ('Quota')}

\item{dataYrs}{is a vector of years on which to base recent catches. Used
when  `basis = 'recent_catch'`.}

\item{advYr}{is a vector of the year in which the scenario catches are generated.}

\item{sc}{is a vector with the scenario to plot, e.g. "min"}

\item{fleets_excl}{is a vector of fleet names not to plot, e.g. "OTH_OTH"}

\item{refTable}{data.frame Contains stock look-up information for consistent
plotting of stocks. `Advice_name` defines the stock names corresponding to
`data` object. `col` defines the color used to fill bars in plot.
`order` defines the order of stocks in the plot facets.}

\item{xlab}{character X-axis label (Default: `xlab = "Stock"`)}

\item{ylab}{character Y-axis label (Default: `ylab = "KW days ('000)"`)}

\item{fillLegendTitle}{character Fill legend title
(Default: `fillLegendTitle = "Effort stock"`)}

\item{colLegendTitle}{character Color legend title
(Default: `colLegendTitle = "Limiting stock"`)}
}
\value{
plot output of class ggplot
}
\description{
Plot of a fleets catch difference from the recent catches or the quota.
  By fleet. Most- and least-limiting stocks are also
  denoted. Testing in response to WKMIXFISH2.
}
\details{
Users will need to provide the data and reference table objects to
  produce the plot.
}
\examples{

# make example data
data(refTable) # reference table with stock advice names, colors, order, etc.
data(stfFltStkSum) # summary of fleet/stock-related catch variables
advYr <- 2022 # advice year

plot_catch_change(data = stfFltStkSum,
 basis = "Quota",
 dataYrs = 2020:2022,
 advYr = advYr,
 sc = "min",
 fleets_excl = "OTH_OTH",
 refTable = refTable,
 xlab = "Stock",
 ylab = "landings change (tonnes)",
 fillLegendTitle = "Stock",
 colLegendTitle = "Limiting stock")

}
