% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catchAlluvial.R
\name{plot_catchAlluvial}
\alias{plot_catchAlluvial}
\title{Alluvial plot of catch}
\usage{
plot_catchAlluvial(
  data,
  refTable,
  text_repel = FALSE,
  text_size = 2,
  xlab = NULL,
  ylab = "Catch [t]",
  fillLegendTitle = "Stock",
  addLegend = TRUE,
  plotTitle = NULL
)
}
\arguments{
\item{data}{data.frame Contains information on catch (or landings)
corresponding to different fleet, metier, and stock combinations.
Variable names should include `stock`, `fleet`, `metier`, and `value`.
Stock variable levels (`stock`) should match those found in `refTable`
containing  color and order information.}

\item{refTable}{data.frame Contains stock look-up information for consistent
plotting of stocks. `stock` defines the stock names corresponding to
`data` object. `col` defines the color used to fill bars in plot.
`order` defines the order of stocks in the plot facets.}

\item{text_repel}{logical Should stratum labels  (`stock`, `fleet`, `metier`)
be repelled to prevent overlapping (using `ggrepel::geom_text_repel`)
(Default: `text_repel = FALSE`)}

\item{text_size}{numeric Value for text label size
(Default: `text_size = 2`)}

\item{xlab}{character X-axis label (Default: `xlab = NULL` removes label)}

\item{ylab}{character Y-axis label (Default: `ylab = Catch [t]`)}

\item{fillLegendTitle}{character Fill legend title
(Default: `fillLegendTitle = "Stock"`)}

\item{addLegend}{logical Should legend of stock fill colors be added}

\item{plotTitle}{character Plot title
(Default: `plotTitle = NULL` removes title)}
}
\value{
plot output of class ggplot
}
\description{
Alluvial plot of catch
}
\examples{

data("stfMtStkSum")
data("refTable")

df <- subset(stfMtStkSum, year == 2020 & scenario == "min" &
  fleet \%in\% unique(stfMtStkSum$fleet)[1:5])[,c("fleet", "metier",
  "stock", "landings")]
df$stock <- refTable$stock[match(df$stock, refTable$stock_short)]
names(df)[4] <- "value"

plot_catchAlluvial(data = df, refTable = refTable)

# repel labels
plot_catchAlluvial(data = df, refTable = refTable,
  text_repel = TRUE, text_size = 2)

# repel labels and suppress legend
plot_catchAlluvial(data = df, refTable = refTable,
  text_repel = TRUE, text_size = 2, addLegend = FALSE)

}
