% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catchCompSpp.R
\name{plot_catchCompSpp}
\alias{plot_catchCompSpp}
\title{Plot landings (catch) composition by country and species}
\usage{
plot_catchCompSpp(data, land_threshold = 10)
}
\arguments{
\item{data}{data.frame Contains information on the total landings of species
listed in the WGMIXFISH data call by country and metier category.
Required variables are: `country` (2 letter country code),
`metCat` (which indicates the metier categories), `MIXFISHspp` (which
indicates the 3 letter FAO species code as listed in the WGMIXFISH
data call), and `landings`.}

\item{land_threshold}{numeric A number, in tonnes, to set a minimum landings
threshold. Metier-country combinations with total landings below this
threshold will not be plotted. Default value is 10 tonnes.}
}
\value{
plot output of class ggplot
}
\description{
landings compositions of species landed by country and metier
  category
}
\details{
Users will need to provide the data to produce the plot.
}
\examples{

# Landings - prepare example data
data(NSlandings)

dat <- NSlandings
# aggregate some metiers
dat <- dat \%>\% mutate(metCat = if_else(metCat \%in\%
  c("dredges","beam_oth"),"MIS",metCat))
dat <- dat \%>\% mutate(
  metCat = if_else(metCat \%in\%
   c("Dseine_TR3","Dseine_TR1","Dseine_TR2","Dseine_oth"),
   "Demersal seine", metCat))

# re-aggregate
datayear <- 2023
dat <- dat \%>\% filter(Year \%in\% datayear) \%>\%
  group_by(country=Country,metCat,MIXFISHspp) \%>\%
  summarise(landings=sum(Landings))

# Set landings threshold
land_th <- 10 # 10 tonnes.

# plot
# windows()
plot_catchCompSpp(dat, land_threshold=land_th)

}
