% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catchScenStk.R
\name{plot_catchScenStk}
\alias{plot_catchScenStk}
\title{Headline advice plot}
\usage{
plot_catchScenStk(data, xlab = "Scenario", ylab = "Catch [t]")
}
\arguments{
\item{data}{data.frame Contains catch (`catch`) and quota (`quota`) levels
by scenario (`catch`) and stock (`stock`).}

\item{xlab}{character X-axis label}

\item{ylab}{character Y-axis label}
}
\value{
plot output of class ggplot
}
\description{
Plot summarizing over- and under-quota catches by stock and
  scenario. Dashed line displays quota by stock. Colored background further
  emphasizes over- and under-quota catches (under - green, over - red).
}
\examples{

data("stfFltStkSum")
head(stfFltStkSum)

df <- stfFltStkSum
advYr <- 2022 # advice year
restr.stks <- c("COD-NS", "HAD", "PLE-EC", "PLE-NS", "POK", "SOL-EC",
  "SOL-NS", "TUR", "WHG-NS", "WIT", "NEP6", "NEP7", "NEP8", "NEP9")

df <- subset(df, year == advYr & stock \%in\% restr.stks)
tmp1 <- aggregate(catch ~ scenario + stock, data = df, FUN = sum)
tmp2 <- aggregate(quota ~ stock, data = subset(df, scenario == "min"),
  FUN = sum)
catchScenStk <- merge(x = tmp1, y = tmp2, all = TRUE)

# re-label scenarios
catchScenStk$scenario <- factor(catchScenStk$scenario,
  levels = c("sq_E", "max", "min", "cod-ns"),
  labels = c("Sq-E", "Max", "Min", "COD-NS"))
head(catchScenStk)

p <- plot_catchScenStk(data = catchScenStk)
print(p)

# png("catchScenStk1.png", width = 6, height = 5, units = "in", res = 400)
# print(p); dev.off()

}
