% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FOgearType.R
\name{plot_FOgearType}
\alias{plot_FOgearType}
\title{Plot total landings or effort by broad gear categories}
\usage{
plot_FOgearType(data)
}
\arguments{
\item{data}{data.frame Contains information on the total landings or effort
by year and gear categories. Required variables are: `year`, `gearType`
(which indicates the broad gear categories), `value` (which indicates
the total to be plotted), and `dataType` which indicates the type of data
to be plotted (either \code{"landings"}, \code{"days at sea"} or
\code{"KW days"}.}
}
\value{
plot output of class ggplot
}
\description{
total landings or effort by broad gear categories such as 'otter trawls', ' beam trawls'
}
\details{
Users will need to provide the data to produce the plot.
}
\examples{

# Landings - prepare example data
data(NSlandings)

dat <- NSlandings
dat <- rename(dat,gearType=gearCat)
dat <- rename(dat,year=Year)

# aggregate landings data
dat <- dat  \%>\% group_by(year,gearType) \%>\%
  summarise(value=sum(Landings,na.rm=TRUE))
dat$dataType <- "landings"

# plot
# windows()
plot_FOgearType(dat)


# Effort - prepare example data
data(NSeffort)

dat <- NSeffort
dat <- rename(dat,gearType=gearCat)
dat <- rename(dat,year=Year)
dat <- dat  \%>\% group_by(year,gearType) \%>\%
  summarise(value=sum(Days_at_sea,na.rm=TRUE))
dat$dataType <- "days at sea"

# plot
# windows()
plot_FOgearType(dat)


}
