% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MetierSankey.R
\name{plot_MetierSankey}
\alias{plot_MetierSankey}
\title{Metier Sankey plot}
\usage{
plot_MetierSankey(Data, Col_2_Link = NA)
}
\arguments{
\item{Data}{data.frame containing the original metier from the accession file
the output metier, to be used in the model and a Link value (default assumption
is Landings)}

\item{Col_2_Link}{column name (character) for the linking "value" variable. Default assumption is
NA and the function defaults to Landings column}
}
\value{
a sanky plot, see the example for how to save a static sankey plot.
}
\description{
function to plot metier to mixedfish metier flow
to provide a description and visualization of how metiers are constructed
}
\details{
Users will need to provide a data frame with three columns, two for metiers
and one for the value used to link them. The data must be surmised to the metier columns
using a group_by statement or similar. Where a metier goes to itself for example SDN_DEF
to SDN_DEF you will experiences a doughnut
}
\examples{
library(tidyverse)
library(networkD3)
library(htmlwidgets)

Dat_4_plot <- Dat_4_sanky \%>\% filter(Country == "UKE") \%>\% group_by_at(vars(-Year,-Landings,-Stock)) \%>\% summarise(Landings=sum(Landings)) \%>\% ungroup()

P <- plot_MetierSankey(Data, Col_2_Link=="Effort")

Sankey plots are interactive by nature and are saved as an html, to get a static image they
are captured using webshot from the htmlwidgets

P <- htmlwidgets::prependContent(P, htmltools::tags$h1("Title"))
# save plot


}
