% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overUnderFltStk.R
\name{plot_overUnderFltStk}
\alias{plot_overUnderFltStk}
\title{Plot over- and undershoot of stock quotas by fleet}
\usage{
plot_overUnderFltStk(
  data,
  refTable,
  yExt = 0.3,
  xlab = "Stock",
  ylab = "Predicted catch [t] with advice undershoot (negative extent)",
  borderSize = 0.5,
  fillLegendTitle = "Stock",
  colLegendTitle = "Limiting stock"
)
}
\arguments{
\item{data}{data.frame Contains information on effort required to uptake
quotas by fleet and stock, plus designation of each stock's limitation
status to the fleet's fishing effort. Stock variable names (`Advice_name`)
should match those of \code{\link[mixfishtools]{refTable}}.
Other required variables include: `Limitation` - defines, by fleet, the
most- (`most`), least- (`least`), and intermediate-limiting (`NA`)
stocks; `quotaEffort` - the effort, by fleet, required to take up the
quota share of each stock; `sqEffort` - status quo effort corresponding
to most recent data year before forecast.}

\item{refTable}{data.frame Contains stock look-up information for consistent
plotting of stocks. `Advice_name` defines the stock names corresponding to
`data` object. `col` defines the color used to fill bars in plot.
`order` defines the order of stocks in the plot facets.}

\item{yExt}{Fraction of absolute range to extend y-axis for each fleet
facet (Default: yExt = 0.3).}

\item{xlab}{character X-axis label (Default: xlab = "Stock")}

\item{ylab}{character Y-axis label (Default: ylab = "Predicted catch [t]
with advice undershoot (negative extent)")}

\item{borderSize}{line width of border around bars (Default: borderSize=0.5)}

\item{fillLegendTitle}{character Fill legend title
(Default: `fillLegendTitle = "Stock"`)}

\item{colLegendTitle}{character Color legend title
(Default: `colLegendTitle = "Limiting stock"`)}
}
\value{
plot output of class ggplot
}
\description{
Plot of over- and undershoot of each stock's quota
  by fleet. Most- and least-limiting stocks are also
  denoted.
}
\details{
Users will need to provide the data and reference table objects to
  produce the plot.
  In the best case, effort associated with complete quota uptake by
  fleet (`data$effortQuota`) may be derived from scenarios restricting fleet
  catch one stock at a time. In the following example, however, effort
  levels are derived by linearly extrapolating the quota uptake levels
  by the effort of the "min" scenario. This is strictly linear when quotas
  are based on partial F, as in FCube. In FLBEIA, quotas are based on catch
  (or landings), which may deviate from a linear relationship when a stock
  is close full exploitation (should not result from an ICES harvest control
  rule).
}
\examples{

# example data for plot_effortFltStk --------------------------------------

data(refTable) # reference table with stock advice names, colors, order, etc.
data(stfFltSum) # summary of fleet-related variables (e.g. effort)
data(stfFltStkSum) # summary of fleet/stock-related catch variables

## get data from advice year

# catches by fleet and stock
advYr <- 2022 # advice year
df <- subset(stfFltStkSum, scenario == "min" & year == advYr)

## effort by fleet and scenario
eff <- subset(
 stfFltSum, scenario == "min" & year == advYr)[,c("fleet", "effort")]
sqEff <- subset(
 stfFltSum, scenario == "sq_E" & year == advYr)[,c("fleet", "effort")]
names(sqEff)[2] <- "sqEffort"
eff <- merge(x = eff, y = sqEff, all.x = TRUE)
df <- merge(x = df, y = eff, all.x = TRUE)
df$quotaEffort <- df$effort / df$quotaUpt


## Determine most- and least-limiting stock by fleet
# restrictive stocks
restr.stks <- c("COD-NS", "HAD", "PLE-EC", "PLE-NS", "POK", "SOL-EC",
 "SOL-NS", "TUR", "WHG-NS", "WIT", "NEP6", "NEP7", "NEP8", "NEP9")
fls <- unique(df$fleet)
df2 <- vector("list", length(fls))
names(df2) <- fls
for(i in seq(fls)){
 tmp <- subset(df, fleet == fls[i])
 tmp$Limitation <- NA # initial NA setting for all stocks

 # most-limiting (highest quota uptake in min scenario)
 mostLimStk <- subset(tmp, stock \%in\% restr.stks)
 mostLimStk <- mostLimStk$stock[which.max(mostLimStk$quotaUpt)]
 tmp$Limitation[which(tmp$stock == mostLimStk)] <- "most"

 # least-limiting (lowest quota uptake in max scenario)
 leastLimStk <- subset(stfFltStkSum, scenario == "max" & year == advYr &
   fleet == fls[i] & stock \%in\% restr.stks)
 leastLimStk <- leastLimStk$stock[which.min(leastLimStk$quotaUpt)]
 tmp$Limitation[which(tmp$stock == leastLimStk)] <- "least"

 # return result
 df2[[i]] <- tmp
}
df2 <- do.call("rbind", df2)

# replace short stock names with ICES stock codes
df2$stock <- refTable$stock[match(df2$stock, refTable$stock_short)]


# plot
p <- plot_overUnderFltStk(data = df2, refTable = refTable)
# png("overUnderFltStk1.png", width = 8, height = 10, units = "in", res = 400)
# print(p); dev.off()

# adjust ggplot2 settings
p <- p + theme(text = element_text(size = 12))
# png("overUnderFltStk2.png", width = 8, height = 10, units = "in", res = 400)
# print(p); dev.off()


}
