% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stkMetProp.R
\name{plot_stkMetProp}
\alias{plot_stkMetProp}
\title{Plot proportion of stock landings by metier}
\usage{
plot_stkMetProp(data, text_size = 9, face = "plain")
}
\arguments{
\item{data}{data.frame Contains information on the stock, metiers and
landings proportions. Required variables are: `Stock` (numeric; numbers
the stocks in alphabetical order), `metCat` (which indicates the metier
categories), `prop_stock` (which indicates proportion of total stock
landings associated with each metier category), and `label` (which
indicates the ICES stock code).}

\item{text_size}{size of text}

\item{face}{an aesthetic or parameter inside element_text() that controls
the font style of text elements like titles, axis labels, and legends.}
}
\value{
plot output of class ggplot
}
\description{
landings proportions of each stock associated with different metiers
}
\details{
Users will need to provide the data to produce the plot.
}
\examples{

# Landings - prepare example data
data(stock_props)
# plot
plot_stkMetProp(stock_props)

}
