% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refTable.R
\docType{data}
\name{refTable}
\alias{refTable}
\title{Look-up reference table for stocks and associated attributes}
\format{
bla bla
}
\source{
WGMIXFISH-Advice 2021, North Sea case study.
  (https://github.com/ices-taf/2021_NrS_MixedFisheriesAdvice)
}
\usage{
data(refTable)
}
\description{
The \code{refTable} data.frame lists stock names and
  corresponding colors for consistency across plots. To be used as a look-up
  table in converting between variable stock names and printed ones.

\itemize{
  \item 1) stock - stock names used in mixed fishery model
  \item 2) lab - alternate stock common name labels
  \item 3) ref - number reference
  \item 4) order - stock order to be used in plots
  \item 5) col - stock colors for plots (e.g. pals::brewer.paired())
  \item 6) Advice_name - alternate stock definition name
}
}
\examples{

data(refTable)
refTable


}
\keyword{datasets}
