% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfFltStkSum.R
\docType{data}
\name{stfFltStkSum}
\alias{stfFltStkSum}
\title{Data.frame containing short-term forecast summary of catch-related
  variables per stock and fleet combination}
\format{
data.frame
}
\source{
WGMIXFISH-Advice 2021, North Sea case study.
  \url{[https://github.com/ices-taf/2021_NrS_MixedFisheriesAdvice]}
}
\usage{
data(strFltStkSum)
}
\description{
The \code{stfFltStkSum} data.frame is an output of
  `FLBEIA::fltStkSum()`. Provides example data for use in
  `plot_effortFltStk`.

\itemize{
  \item scenario - advice year
  \item year - advice year
  \item fleet - fleet names
  \item stock - stock names used in mixed fishery model
  \item Advice_name - stock code labels
  \item Limitation - by fleet, identifies limiting stocks
    (levels: choke, interm., least). Describes the most-, intermediate-,
    and least-limiting quotas based on quota uptake ratios (quotaUpt).
  \item quotaEffort - effort required to take up quota in advice year
  \item sqE_effort - status quo effort from last data year
  \item rel_effort - ratio of quota_effort / sqE_effort
}
}
\examples{

data(stfFltStkSum)
head(stfFltStkSum)


}
\keyword{datasets}
