% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_props.R
\docType{data}
\name{stock_props}
\alias{stock_props}
\title{Data.frame contains information on the stock, metiers and
  landings proportions}
\format{
data.frame
}
\source{
Intercatch
}
\usage{
data(stock_props)
}
\description{
The \code{stock_props} data.frame provides example data for
  use in `plot_stkMetProp`.

\itemize{
  \item Stock - scenario name
  \item metCat - year
  \item prop_stock - fleet names
  \item label
}
}
\examples{

data(stock_props)
head(stock_props)


}
\keyword{datasets}
