% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgnsskStocks.R
\docType{data}
\name{wgnsskStocks}
\alias{wgnsskStocks}
\title{FLStocks object with assessment estimated values}
\format{
FLStocks
}
\source{
WGNSSK 2024
}
\usage{
data(wgnsskStocks)
}
\description{
The \code{wgnsskStocks} data is an FLStocks object containing
  estimated values for 6 stocks assessed during WGNSSK 2024
  ("HAD", "NEP6", "PLE-EC", "PLE-NS", "POK", "WIT").
}
\examples{

library(FLCore)
data(wgnsskStocks)
wgnsskStocks[["HAD"]]


}
\keyword{datasets}
