% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catchScenStk.R
\name{plot_catchScenStk}
\alias{plot_catchScenStk}
\title{Headline advice plot}
\usage{
plot_catchScenStk(
  data,
  adv,
  ofwhich = FALSE,
  xlab = "Scenario",
  ylab = "Catch [t]"
)
}
\arguments{
\item{data}{data.frame Contains catch (`catch`) by scenario (`scenario`) and
stock (`stock`).}

\item{adv}{data.frame Contains advice (`advice`) by stock (`stock`). Optional
upper (`stock`) and lower (`lower`) advice limits can be included.}

\item{ofwhich}{logical. If TRUE an 'off which limit' will be plotted.
Requires a 'catch_ofwhich' column in data and an 'advice_ofwhich' column in adv.}

\item{xlab}{character X-axis label (Default: `xlab = "Scenario"`)}

\item{ylab}{character Y-axis label (Default: `ylab = "Catch [t]"`)}
}
\value{
plot output of class ggplot
}
\description{
Plot summarizing over- and under-quota catches by stock and
  scenario. Dashed line displays quota by stock. Colored background further
  emphasizes over- and under-quota catches. Used as the headline plot in
  WGMIXFISH-ADVICE.
}
\examples{

# make example data
data("stfFltStkSum")
head(stfFltStkSum)

# subset data to advice year and restrictive stocks
advYr <- 2022 # advice year
restr.stks <- c("COD-NS", "HAD", "PLE-EC", "PLE-NS", "POK", "SOL-EC",
  "SOL-NS", "TUR", "WHG-NS", "WIT")
stfFltStkSum <- subset(stfFltStkSum, year == advYr & stock \%in\% restr.stks)

# data for plotting (catch by scenario and stock)
catchScenStk <- aggregate(catch ~ scenario + stock, data = stfFltStkSum, FUN = sum)

# re-order scenarios (sq_E, max, min, ... )
catchScenStk$scenario <- factor(catchScenStk$scenario,
  levels = c("sq_E", "max", "min", "cod-ns"),
  labels = c("sq_E", "max", "min", "cod-ns"))
head(catchScenStk)

catchRange <- rbind(
  data.frame(stock = "COD-NS", advice = 14276, lower = 9701, upper = 14276),
  data.frame(stock = "HAD", advice = 128708, lower = 111702, upper = 128708),
  data.frame(stock = "PLE-EC", advice = 6365, lower = 4594, upper = 6365),
  data.frame(stock = "PLE-NS", advice = 142507, lower = 101854, upper = 195622),
  data.frame(stock = "POK", advice = 49614, lower = 30204, upper = 49614),
  data.frame(stock = "SOL-EC", advice = 1810, lower = 1068, upper = 2069),
  data.frame(stock = "SOL-NS", advice = 15330, lower = 9523, upper = 21805),
  data.frame(stock = "TUR", advice = 3609, lower = 2634, upper = 4564),
  data.frame(stock = "WHG-NS", advice = 88426, lower = 70169, upper = 91703),
  data.frame(stock = "WIT", advice = 1206, lower = 875, upper = 1206)
)

# plot without range
p <- plot_catchScenStk(data = catchScenStk, adv = catchRange[,1:2])
print(p)

# plot with range
p <- plot_catchScenStk(data = catchScenStk, adv = catchRange)
print(p)

# export plot
# png("catchScenStk1.png", width = 6, height = 5, units = "in", res = 400)
# print(p); dev.off()

}
