% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effortFltStk.R
\name{plot_effortFltStk}
\alias{plot_effortFltStk}
\title{Fleet effort to uptake stock quotas}
\usage{
plot_effortFltStk(data, refTable)
}
\arguments{
\item{data}{data.frame Contains information on effort required to uptake
quotas by fleet and stock. Variable names should match those of
\code{\link[mixfishtools]{effortFltStk}}
(i.e. stock, Advice_name, Limitation, effortQuota, sqE_effort).}

\item{refTable}{data.frame Contains stock look-up information for consistent
plotting of stocks. Variable names should match those of
\code{\link[mixfishtools]{refTable}}.}
}
\value{
plot output of class ggplot
}
\description{
Fleet effort to uptake stock quotas
}
\examples{
data(refTable)
data(effortFltStk)
p <- plot_effortFltStk(data = effortFltStk, refTable = refTable)
print(p)


}
